fix: move MODEL_LIST to modal dialog
