+++ public/components/monitoring/model_status_filter.tsx
@@ -27,7 +27,6 @@ export const ModelStatusFilter = ({ selection = [], onChange }: Props) => {
     (newOptions: IOption[]) => {
       onChange(newOptions.filter(({ checked }) => checked === 'on').map(({ value }) => value));
     },
-
     [onChange]
   );
